KAS-3106 REwork the new flow for requesting & uploading translations
