--- app/components/publications/publication/translations/documents-edit-modal.hbs
@@ -1,103 +1,0 @@
-<Auk::Modal>
-  <Auk::Modal::Header
-    @title={{t
-      "edit-translation-document"
-      docname=this.name
-      identifier=@translationSubcase.publicationFlow.identification.idName
-    }}
-    @onClose={{@onCancel}}
-  />
-  <Auk::Modal::Body>
-    <div class="auk-o-grid">
-      <div class="auk-o-grid-col-11">
-        <div class="auk-form-group-layout auk-form-group-layout--horizontal">
-          <div class="auk-form-group auk-o-flex--align-start">
-            <Auk::Label for="nameDocumentUpload" @isRequired={{true}}>
-              {{t "name"}}
-            </Auk::Label>
-            <Auk::Input
-              data-test-publication-document-edit-document-name
-              id="nameDocumentUpload"
-              @value={{mut this.name}}
-            />
-          </div>
-          <div class="auk-form-group">
-            <Auk::Label>{{t "document-type"}}</Auk::Label>
-            <Auk::Pill @skin="white">
-              {{@sourceDocument.file.extension}}
-            </Auk::Pill>
-          </div>
-          <div class="auk-form-group auk-o-flex--align-start">
-            <Auk::Label>{{t "translation-burden"}}</Auk::Label>
-            <div
-              class="auk-form-group-layout auk-form-group-layout--standard auk-form-group-layout--grid"
-            >
-              <div class="auk-o-grid">
-                <div class="auk-o-grid-col-6">
-                  <div class="auk-form-group">
-                    <Auk::Label
-                      for="translationLoadPages"
-                      class="auk-label--light"
-                    >
-                      {{t "pages-amount"}}
-                    </Auk::Label>
-                    <Auk::Input
-                      data-test-publication-document-edit-pages
-                      @type="number"
-                      id="translationLoadPages"
-                      @value={{mut this.pagesAmount}}
-                    />
-                  </div>
-                </div>
-                <div class="auk-o-grid-col-6">
-                  <div class="auk-form-group">
-                    <Auk::Label
-                      for="translationLoadWords"
-                      class="auk-label--light"
-                    >
-                      {{t "words-amount"}}
-                    </Auk::Label>
-                    <Auk::Input
-                      data-test-publication-document-edit-words
-                      @type="number"
-                      id="translationLoadWords"
-                      @value={{mut this.wordsAmount}}
-                    />
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div class="auk-form-group auk-o-flex--align-start">
-            <Auk::Label for="proofPrint">
-              {{t "for-proof-print"}}
-            </Auk::Label>
-            {{#if this.loadData.isRunning}}
-              <Auk::Loader @message={{t "loading"}} />
-            {{else}}
-              <Auk::CheckboxList>
-                <Auk::Checkbox
-                  data-test-publication-document-edit-proofprint
-                  @name="proofPrint"
-                  @checked={{this.isSourceForProofPrint}}
-                  {{on "input" this.toggleProofprint}}
-                />
-              </Auk::CheckboxList>
-            {{/if}}
-          </div>
-        </div>
-      </div>
-    </div>
-  </Auk::Modal::Body>
-  <Auk::Modal::Footer @onCancel={{@onCancel}}>
-    <Auk::Button
-      data-test-publication-document-edit-save
-      @skin="primary"
-      @disabled={{this.saveIsDisabled}}
-      @loading={{this.saveEditDocument.isRunning}}
-      {{on "click" (perform this.saveEditDocument)}}
-    >
-      {{t "save"}}
-    </Auk::Button>
-  </Auk::Modal::Footer>
-</Auk::Modal>
