--- app/components/publications/publication/translations/documents-upload-modal.js
@@ -1,116 +1,0 @@
-import Component from '@glimmer/component';
-import { action } from '@ember/object';
-import { inject as service } from '@ember/service';
-import { isPresent } from '@ember/utils';
-import { tracked } from '@glimmer/tracking';
-import { task } from 'ember-concurrency-decorators';
-import { guidFor } from '@ember/object/internals';
-import {
-  ValidatorSet, Validator
-} from 'frontend-kaleidos/utils/validators';
-
-export default class PublicationsTranslationDocumentUploadModalComponent extends Component {
-  /**
-   * @argument onSave: should take arguments (pieces)
-   * @argument onCancel
-   */
-  @service store;
-  @service('file-queue') fileQueueService;
-
-  @tracked translationDocument;
-  @tracked name;
-  @tracked pagesAmount;
-  @tracked wordsAmount;
-  @tracked isSourceForProofPrint = false;
-
-  validators;
-
-  constructor() {
-    super(...arguments);
-    if (this.fileQueueService.find(this.fileQueueName)) {
-      this.fileQueueService.create(this.fileQueueName);
-    }
-    this.initValidators();
-  }
-
-  get fileQueueName() {
-    return `${guidFor(this)}-file-queue`;
-  }
-
-  get fileQueue() {
-    return this.fileQueueService.find(this.fileQueueName);
-  }
-
-  get isCancelDisabled() {
-    return this.cancel.isRunning || this.save.isRunning;
-  }
-
-  get isSaveDisabled() {
-    return !this.translationDocument || !this.validators.areValid || this.cancel.isRunning;
-  }
-
-  @action
-  uploadPiece(file) {
-    const now = new Date();
-    const documentContainer = this.store.createRecord('document-container', {
-      created: now,
-    });
-    this.translationDocument = this.store.createRecord('piece', {
-      created: now,
-      modified: now,
-      file: file,
-      confidential: false,
-      name: file.filenameWithoutExtension,
-      documentContainer: documentContainer,
-    });
-    this.name = file.filenameWithoutExtension;
-  }
-
-  @task({
-    drop: true,
-  })
-  *cancel() {
-    // necessary because close-button is not disabled when saving
-    if (this.save.isRunning) {
-      return;
-    }
-
-    if (this.translationDocument) {
-      yield this.deleteUploadedPiece.perform(this.translationDocument);
-    }
-    this.args.onCancel();
-  }
-
-  @task
-  *deleteUploadedPiece(piece) {
-    const file = yield piece.file;
-    yield file.destroyRecord();
-    const documentContainer = yield piece.documentContainer;
-    yield documentContainer.destroyRecord();
-    yield piece.destroyRecord();
-  }
-
-  @task
-  *save() {
-    if (this.args.onSave) {
-      yield this.args.onSave({
-        piece: this.translationDocument,
-        name: this.name,
-        pagesAmount: this.pagesAmount,
-        wordsAmount: this.wordsAmount,
-        isSourceForProofPrint: this.isSourceForProofPrint,
-      });
-    }
-  }
-
-  @action
-  toggleProofprint() {
-    this.isSourceForProofPrint = !this.isSourceForProofPrint;
-  }
-
-  initValidators() {
-    this.validators = new ValidatorSet({
-      name: new Validator(() => isPresent(this.name)),
-    });
-  }
-}
