+++ app/components/publications/publication/translations/translation-upload-modal.hbs
@@ -37,12 +37,6 @@
                     {{on "input" this.validators.name.enableError}}
                   />
                 </div>
-                <div class="auk-form-group">
-                  <Auk::Label>{{t "file-type"}}</Auk::Label>
-                  <Utils::Documents::FileTypePill
-                    @extension={{this.file.extension}}
-                  />
-                </div>
                 <div
                   class="auk-form-group
                     {{this.validators.receivedAtDate.errorClass}}"
@@ -54,16 +48,6 @@
                     id="translationUploadReceivedAtDate"
                     @date={{this.receivedAtDate}}
                     @onChange={{this.setReceivedAtDate}}
-                  />
-                </div>
-                <div class="auk-form-group auk-o-flex--align-start">
-                  <Auk::Label for="translationUploadForProofPrint">
-                    {{t "for-proof-print"}}
-                  </Auk::Label>
-                  <Auk::Checkbox
-                    id="translationUploadForProofPrint"
-                    @checked={{this.isSourceForProofPrint}}
-                    {{on "input" this.toggleProofprint}}
                   />
                 </div>
                 <div class="auk-form-group auk-o-flex--align-start">
